Update for search_engine_icon_in_searchbar.uc.js
Firefox 77 removed the call to setIcon() that this was hooking. We can
hook updateDisplay() instead, which should probably still work with
earlier versions too.

Also, I've noticed that the script doesn't like to properly run. It
turns out that the 1000ms delay at the start is sometimes not enough, so
let's catch the exception and retry a few times.

Fixes Aris-t2/CustomJSforFx#33